Lesson 173
Using rxjs interval observable which emits a count each X milisecond
We also stored it in a subscription
Then we unsubscribed each time we left the component to prevent memory leaks!
We also learned that if we don't unsubscribe, we create a new observable on top of we already have.