Lesson 176 - Observables & You!
We usually use observables from libs. We probably won't use custom observables that often.

this.route.params.subscribe( ( params:Params) => {
this.id = +params.id
})

In the previous example we used route.params observable which is an angular built in observable. Therefore we don't need to unsubscribed it too! (Unlike rxjs observables!)

When we subscribe we will pass the functions that deal with passed data or errors.

 :gem: Ex :
customNewObsSubscription2  sends customIntervalObservable3
3 types of functions :
      dat =>
      {
        console.log(dat);
      }
the normal case (Continues as usual)

      error => {
        console.log(error);
        alert(error.message)
      }
error case (Halts) and,
      () => {
        console.log('Completed!')
      }
Completed case. (Halts)