Lesson 178 - Subjects
Use Subjects instead of EventEmitters!
But only if there are no @Outputs in the component.
You only use subjects to commmunicate across components, through services so through a mechanism where you in the end subscribe to somewhere, like here in the app component.

If you're not subscribing to an event emitter, then it probably is an output
If you do plan to subscribe manually, then it is a subject.